control_room/requirements.txt
streamlit
opencv-python
numpy
pyaudio
librosa
ultralytics
pymongo
bcrypt
python-socketio
eventlet
pyserial
